Refactor hardhat config to have separate defaults file and have optional upgrade tasks
