+++ packages/arb-bridge-eth/hardhat.default-config.ts
@@ -1,0 +1,198 @@
+import { task } from 'hardhat/config'
+import '@nomiclabs/hardhat-ethers'
+import 'hardhat-deploy'
+import 'dotenv/config'
+import '@nomiclabs/hardhat-waffle'
+import '@typechain/hardhat'
+import 'solidity-coverage'
+import 'hardhat-spdx-license-identifier'
+import 'hardhat-gas-reporter'
+import '@nomiclabs/hardhat-etherscan'
+
+task('accounts', 'Prints the list of accounts', async (taskArgs, bre) => {
+  const accounts = await bre.ethers.getSigners()
+
+  for (const account of accounts) {
+    console.log(await account.getAddress())
+  }
+})
+
+const config = {
+  defaultNetwork: 'hardhat',
+  paths: {
+    artifacts: 'build/contracts',
+  },
+  typechain: {
+    outDir: 'build/types',
+    target: 'ethers-v5',
+  },
+  spdxLicenseIdentifier: {
+    overwrite: false,
+    runOnCompile: true,
+  },
+  gasReporter: {
+    currency: 'USD',
+    gasPrice: 20,
+    enabled: process.env.REPORT_GAS ? true : false,
+  },
+  namedAccounts: {
+    deployer: {
+      default: 0,
+    },
+  },
+  networks: {
+    hardhat: {
+      chainId: 1337,
+      throwOnTransactionFailures: true,
+      allowUnlimitedContractSize: true,
+      accounts: {
+        accountsBalance: '1000000000000000000000000000',
+      },
+      blockGasLimit: 200000000,
+      // mining: {
+      //   auto: false,
+      //   interval: 1000,
+      // },
+      forking: {
+        url: 'https://mainnet.infura.io/v3/' + process.env['INFURA_KEY'],
+        enabled: process.env['SHOULD_FORK'] === '1',
+      },
+    },
+    local_development: {
+      url: 'http://127.0.0.1:7545',
+    },
+    kovan: {
+      url: 'https://kovan.infura.io/v3/' + process.env['INFURA_KEY'],
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    mainnet: {
+      url: 'https://mainnet.infura.io/v3/' + process.env['INFURA_KEY'],
+      accounts: process.env['MAINNET_PRIVKEY']
+        ? [process.env['MAINNET_PRIVKEY']]
+        : [],
+    },
+    fork: {
+      url: 'http://127.0.0.1:8545/',
+    },
+    arbitrum1: {
+      url: 'https://arb1.arbitrum.io/rpc',
+      accounts: process.env['MAINNET_PRIVKEY']
+        ? [process.env['MAINNET_PRIVKEY']]
+        : [],
+    },
+    rinkeby: {
+      url: 'https://rinkeby.infura.io/v3/' + process.env['INFURA_KEY'],
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    arbRinkeby: {
+      gasPrice: 0,
+      url: 'https://rinkeby.arbitrum.io/rpc',
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    arbkovan4: {
+      gasPrice: 0,
+      url: 'https://kovan4.arbitrum.io/rpc',
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    kovan5: {
+      gasPrice: 0,
+      url: 'https://kovan5.arbitrum.io/rpc',
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    devnet: {
+      url: 'https://devnet.arbitrum.io/rpc',
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    devnetL2: {
+      url: 'https://devnet-l2.arbitrum.io/rpc',
+      accounts: process.env['DEVNET_PRIVKEY']
+        ? [process.env['DEVNET_PRIVKEY']]
+        : [],
+    },
+    arbitrum: {
+      url: 'http://127.0.0.1:8547',
+      // url: 'https://kovan3.arbitrum.io/rpc',
+      gas: 999999999999999,
+      accounts: {
+        mnemonic:
+          'jar deny prosper gasp flush glass core corn alarm treat leg smart',
+        path: "m/44'/60'/0'/0",
+        initialIndex: 0,
+        count: 10,
+      },
+      timeout: 100000,
+    },
+  },
+  mocha: {
+    timeout: 0,
+    bail: true,
+  },
+  etherscan: {
+    apiKey: process.env['ETHERSCAN_API_KEY'],
+  },
+  solidity: {
+    compilers: [
+      {
+        version: '0.6.11',
+        settings: {
+          optimizer: {
+            enabled: true,
+            runs: 100,
+          },
+        },
+      },
+      {
+        version: '0.8.7',
+        settings: {
+          optimizer: {
+            enabled: true,
+            runs: 100,
+          },
+        },
+      },
+    ],
+  },
+}
+
+if (process.env['RINKEBY_URL'] && process.env['RINKEBY_MNEMONIC']) {
+  ;(config.networks as any)['rinkeby'] = {
+    url: process.env['RINKEBY_URL'] || '',
+    accounts: [process.env['RINKEBY_MNEMONIC'] || ''],
+    network_id: 4,
+    confirmations: 1,
+  }
+}
+
+if (process.env['ROPSTEN_URL'] && process.env['ROPSTEN_MNEMONIC']) {
+  ;(config.networks as any)['ropsten'] = {
+    url: process.env['ROPSTEN_URL'] || '',
+    accounts: [process.env['ROPSTEN_MNEMONIC'] || ''],
+    network_id: 3,
+    confirmations: 1,
+  }
+}
+
+if (process.env['KOVAN_URL'] && process.env['KOVAN_MNEMONIC']) {
+  ;(config.networks as any)['kovan'] = {
+    url: process.env['KOVAN_URL'] || '',
+    accounts: [process.env['KOVAN_MNEMONIC'] || ''],
+    network_id: 42,
+    confirmations: 4,
+  }
+}
+
+if (!process.env['DEVNET_PRIVKEY']) console.warn('No devnet privkey set')
+
+export { config }
