+++ packages/arb-bridge-peripherals/hardhat.config.ts
@@ -1,57 +1,9 @@
-import * as config from 'arb-bridge-eth/hardhat.config'
-import { task } from 'hardhat/config'
-import { initUpgrades } from 'arb-upgrades'
-import '@nomiclabs/hardhat-ethers'
+import { config } from 'arb-bridge-eth/hardhat.default-config'
 
-task('deploy-logic-one', 'deploy one logic')
-  .addParam('contract', 'contract to deploy')
-  .setAction(async (args, hre) => {
-    const { contract } = args
-    const { deployLogic } = initUpgrades(hre, __dirname)
-    await deployLogic(contract)
-  })
-
-task('deploy-logic-all', 'deploy all logic contracts').setAction(
-  async (_, hre) => {
-    const { deployLogicAll } = initUpgrades(hre, __dirname)
-    await deployLogicAll()
-  }
-)
-
-task('trigger-upgrades', 'triggers upgrade').setAction(async (_, hre) => {
-  const { updateImplementations } = initUpgrades(hre, __dirname)
-  await updateImplementations()
-})
-
-task('verify-deployments', 'verifies implementations').setAction(
-  async (_, hre) => {
-    const { verifyCurrentImplementations } = initUpgrades(hre, __dirname)
-    await verifyCurrentImplementations()
-  }
-)
-
-task('transfer-owner', 'deploy one logic')
-  .addParam('proxyaddress', 'proxy address')
-  .addParam('newadmin', 'address of new admin')
-  .setAction(async (args, hre) => {
-    const { contract } = args
-    const { transferAdmin } = initUpgrades(hre, __dirname)
-    await transferAdmin(args.proxyaddress, args.newadmin)
-  })
-
-task(
-  'remove-build-info',
-  'remove giant build info string from current_deployments json'
-).setAction(async (_, hre) => {
-  const { removeBuildInfoFiles } = initUpgrades(hre, __dirname)
-  await removeBuildInfoFiles()
-})
-
-task('etherscan-verify', 'verify current deployments in etherscan').setAction(
-  async (_, hre) => {
-    const { verifyDeployments } = await initUpgrades(hre, __dirname)
-    await verifyDeployments()
-  }
-)
+try {
+  require('arb-upgrades/peripheralsTasks')
+} catch (e) {
+  // arb-upgrades dependency not available
+}
 
 module.exports = config
